requirements.txt
PyQt5>=5.15.0
pyautogui>=0.9.53
keyboard>=0.13.5
opencv-python>=4.8.0
Pillow>=10.0.0
pyinstaller>=6.0.0 